backend/requirements.txt
# Core dependencies for RAG chatbot
fastapi>=0.104.1
uvicorn>=0.24.0
python-pptx>=0.6.23
markdown>=3.5.1
beautifulsoup4>=4.12.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=1.3.7
pydantic>=2.5.0
python-multipart>=0.0.6
# numpy and torch are typically already installed with sentence-transformers
# If not, install separately: pip install numpy torch

